# Fix cases where KMetadata interface has nil object