@serathius 2021-03-12
Removed second call to `reflect.ValueOf`, results look much better:
```
name    old time/op    new time/op    delta
KObj-8    42.2ns ± 8%    43.4ns ± 8%  +2.90%  (p=0.000 n=98+98)

name    old alloc/op   new alloc/op   delta
KObj-8     32.0B ± 0%     32.0B ± 0%    ~     (all equal)

name    old allocs/op  new allocs/op  delta
KObj-8      1.00 ± 0%      1.00 ± 0%    ~     (all equal)
```
Looks like the overhead is only 3%!!!