@serathius 2021-03-12
Removing WIP as benchmark results look reasonable