@serathius 2021-03-12
Fixed how benchmark is implemented