@serathius 2021-03-12
Updated benchmark results, looks better around 7% difference.